+++ apps/client/src/app/pages/portfolio/activities/import-activities-dialog/import-activities-dialog.html
@@ -16,10 +16,9 @@
   >
     <mat-step [completed]="importStep === 0" [selected]="importStep === 0">
       <ng-template i18n matStepLabel>Select File</ng-template>
-      <ng-container>
+      <div class="pt-3">
         <ng-container *ngIf="mode === 'DIVIDEND'; else selectFile">
           <form
-            class="mt-3"
             [formGroup]="uniqueAssetForm"
             (ngSubmit)="onLoadDividends(stepper)"
           >
@@ -46,7 +45,7 @@
           </form>
         </ng-container>
         <ng-template #selectFile>
-          <div class="d-flex flex-column justify-content-center mt-3">
+          <div class="d-flex flex-column justify-content-center">
             <button
               class="py-4"
               color="primary"
@@ -74,73 +73,82 @@
             </p>
           </div>
         </ng-template>
-      </ng-container>
+      </div>
     </mat-step>
 
     <mat-step [completed]="importStep === 1" [selected]="importStep === 1">
       <ng-template i18n matStepLabel>Select Activities</ng-template>
-      <ng-container *ngIf="errorMessages.length === 0; else errorMessage">
-        <gf-activities-table
-          *ngIf="importStep === 1"
-          [activities]="activities"
-          [baseCurrency]="data?.user?.settings?.baseCurrency"
-          [deviceType]="data?.deviceType"
-          [hasPermissionToCreateActivity]="false"
-          [hasPermissionToExportActivities]="false"
-          [hasPermissionToFilter]="false"
-          [hasPermissionToOpenDetails]="false"
-          [locale]="data?.user?.settings?.locale"
-          [pageSize]="maxSafeInteger"
-          [showActions]="false"
-          [showCheckbox]="true"
-          [showFooter]="false"
-          [showSymbolColumn]="false"
-          (selectedActivities)="updateSelection($event)"
-        ></gf-activities-table>
-        <div class="d-flex justify-content-end mt-3">
-          <button i18n mat-button (click)="onImportCancel(stepper)">
-            Cancel
-          </button>
-          <button
-            class="ml-1"
-            color="primary"
-            mat-flat-button
-            [disabled]="!selectedActivities?.length"
-            (click)="onImportActivities()"
-          >
-            <ng-container i18n>Import</ng-container>
-          </button>
-        </div>
-      </ng-container>
-      <ng-template #errorMessage>
-        <mat-accordion displayMode="flat">
-          <mat-expansion-panel
-            *ngFor="let message of errorMessages; let i = index"
-            [disabled]="!details[i]"
-          >
-            <mat-expansion-panel-header class="pl-1">
-              <mat-panel-title>
-                <div class="d-flex">
-                  <div class="align-items-center d-flex mr-2">
-                    <ion-icon name="warning-outline"></ion-icon>
+      <div class="pt-3">
+        <ng-container *ngIf="errorMessages.length === 0; else errorMessage">
+          <gf-activities-table
+            *ngIf="importStep === 1"
+            [activities]="activities"
+            [baseCurrency]="data?.user?.settings?.baseCurrency"
+            [deviceType]="data?.deviceType"
+            [hasPermissionToCreateActivity]="false"
+            [hasPermissionToExportActivities]="false"
+            [hasPermissionToFilter]="false"
+            [hasPermissionToOpenDetails]="false"
+            [locale]="data?.user?.settings?.locale"
+            [pageSize]="maxSafeInteger"
+            [showActions]="false"
+            [showCheckbox]="true"
+            [showFooter]="false"
+            [showSymbolColumn]="false"
+            (selectedActivities)="updateSelection($event)"
+          ></gf-activities-table>
+          <div class="d-flex justify-content-end mt-3">
+            <button mat-button (click)="onReset(stepper)">
+              <ng-container i18n>Back</ng-container>
+            </button>
+            <button
+              class="ml-1"
+              color="primary"
+              mat-flat-button
+              [disabled]="!selectedActivities?.length"
+              (click)="onImportActivities()"
+            >
+              <ng-container i18n>Import</ng-container>
+            </button>
+          </div>
+        </ng-container>
+        <ng-template #errorMessage>
+          <mat-accordion displayMode="flat">
+            <mat-expansion-panel
+              *ngFor="let message of errorMessages; let i = index"
+              [disabled]="!details[i]"
+            >
+              <mat-expansion-panel-header class="pl-1">
+                <mat-panel-title>
+                  <div class="d-flex">
+                    <div class="align-items-center d-flex mr-2">
+                      <ion-icon name="warning-outline"></ion-icon>
+                    </div>
+                    <div>{{ message }}</div>
                   </div>
-                  <div>{{ message }}</div>
-                </div>
-              </mat-panel-title>
-            </mat-expansion-panel-header>
-            <pre
-              *ngIf="details[i]"
-              class="m-0"
-            ><code>{{ details[i] | json }}</code></pre>
-          </mat-expansion-panel>
-        </mat-accordion>
-        <div class="mt-2">
-          <button mat-button (click)="onReset(stepper)">
-            <ion-icon class="mr-2" name="arrow-back-outline"></ion-icon>
-            <span i18n>Back</span>
-          </button>
-        </div>
-      </ng-template>
+                </mat-panel-title>
+              </mat-expansion-panel-header>
+              <pre
+                *ngIf="details[i]"
+                class="m-0"
+              ><code>{{ details[i] | json }}</code></pre>
+            </mat-expansion-panel>
+          </mat-accordion>
+          <div class="d-flex justify-content-end mt-3">
+            <button mat-button (click)="onReset(stepper)">
+              <ng-container i18n>Back</ng-container>
+            </button>
+            <button
+              class="ml-1"
+              color="primary"
+              mat-flat-button
+              [disabled]="true"
+            >
+              <ng-container i18n>Import</ng-container>
+            </button>
+          </div>
+        </ng-template>
+      </div>
     </mat-step>
   </mat-stepper>
 </div>
